requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==3.2.11
django-ckeditor==6.2.0
django-js-asset==1.2.2
django-taggit==2.0.0
Pillow==8.4.0
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
